refactor: change structure of. project
