+++ src/utils/util.model.ts
@@ -1,0 +1,13 @@
+export enum ERROR_MSG {
+  USER_ID = 'userId',
+  ARTIST_ID = 'artistId',
+  TRACK_ID = 'trackId',
+}
+
+export enum HTTP_CODE {
+  OK = 200,
+  DELETED = 204,
+  BAD_REQUEST = 400,
+  FORBIDDEN = 403,
+  NOT_FOUND = 404,
+}
